% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICTInternal.R
\name{askUserProceed}
\alias{askUserProceed}
\title{Ask user to confirm whether to proceed}
\usage{
askUserProceed()
}
\value{
A boolean value based on the user's input.
}
\description{
Asks user whether they wish to proceed and prompts user to enter \code{y} or \code{n}.
If the user enters \code{y}, return \code{TRUE}; if the user enters \code{n}, return
\code{FALSE}; if the user enters an invalid input, prompt again.
}
\examples{
print("TODO")
}
