% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareEdgeFeatures.R
\name{calculate_f1}
\alias{calculate_f1}
\title{Replace NA values in a DataFrame based on a replacement list}
\usage{
calculate_f1(results2)
}
\arguments{
\item{df}{A DataFrame in which NA values need to be replaced.}

\item{rplist}{A named list where the names are patterns to match column names in the DataFrame, and the values are the replacement values for NA.}
}
\value{
A DataFrame with NA values replaced according to the specified patterns and replacement values.
}
\description{
This function replaces NA values in a DataFrame with specified values based on a pattern matching of column names.
}
\examples{
df <- data.frame(a1 = c(NA, 2, 3), a2 = c(4, NA, 6), b1 = c(7, 8, NA))
rplist <- list(a = 0, b = 1)
my_replace_na(df, rplist)

}
