% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRawEdges.R
\name{calculateRawEdges}
\alias{calculateRawEdges}
\title{calculateRawEdges}
\usage{
calculateRawEdges(
  gene_expression_matrix = NULL,
  cict_raw_edge_col = "Spearman",
  in_format = "separate",
  n.workers = 5,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{Data.frame of gene expression where rows are
genes and columns are samples}

\item{cict_raw_edge_col}{gene-gene association metric to calculate}

\item{in_format}{String that specifies the input format. Can be 'data_obj',
'separate', or 'config_file'}

\item{n.workers}{Integer. If running parallel, number of threads on which to
run calculations}
}
\value{
Returns a list consisting of three objects
rcrd: is a list object of intermediary objects
edges: a dataframe of edge objects and CICT features for edges
Vertices: a dataframe of vertices objects and CICT features for vertices
}
\description{
Returns a gene-gene association matrix from a gene expression matrix
}
\examples{
c(rcrd,edges,vertices) \%<-z\% prepareEdgeFeatures(Debug=Debug)
}
