% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{Check validity of data}
\usage{
checkData(
  gene_expression_matrix = NULL,
  ground_truth = NULL,
  raw_edges = NULL,
  edge_features = NULL,
  model = NULL,
  model_assessment = NULL,
  predicted_edges = NULL,
  in_data_obj = NULL,
  config_path = NULL,
  in_format = "separate",
  suppress_warnings = FALSE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{The gene expression matrix, where each row
represents a gene and each column represents a sample. A matrix or
DataFrame with numeric data.}

\item{ground_truth}{A DataFrame representing the ground-truth gene regulatory
network for model training and evaluation. Each row represents a
source-target relationship, with the source gene in the column labeled
\code{"src"} and the target gene in the column labeled \code{"trgt"}}

\item{raw_edges}{Edges for the gene association matrix calculated from
expressioon data. Each row represents a source-target relationship, with
the source gene in the column labeled \code{"src"}, the target gene in the
column labeled \code{"trgt"}, and the correlation coefficient in a row named
with \code{Pearson} or \code{Spearman} depending on the type of coefficient used.}

\item{edge_features}{Features calculated from raw edge data to be used in model
prediction}

\item{model}{The model used to predict edges}

\item{model_assessment}{Evaluation of the model used to predict edges}

\item{predicted_edges}{A predicted gene regulatory network.}

\item{in_data_obj}{A list in the CICT data object format. Produced by a CICT
function.}

\item{config_path}{Path to the YAML config file.}

\item{in_format}{String indicating expected input format. \code{"separate"} if
passing inputs as separate arguments, \code{"data_obj"} if passing inputs
through \code{in_data_obj}, \code{"config"} if passing inputs through \code{config_path}.}

\item{...}{Other options}
}
\value{
A list in the CICT data object format, where fields for which valid
data were provided contain the input data and fields for which data
provided was invalid are null.
}
\description{
Checks if inputted data is either valid for use in CICT functions or is a
valid output of a CICT function. Valid data is returned in the CICT data
list. Invalid data has a null field in the list. Does not currently verify
\code{raw_edges}, \code{edge_features}, \code{model}, \code{model_assessment}, or
\code{predicted_edges}.
}
\examples{
# separate inputs
checkData(gene_expression_matrix = SERGIO_DS4_gene_expression_matrix, ground_truth = SERGIO_DS4_ground_truth)

# input a data list
checkData(in_data_obj = checkData(gene_expression_matrix = SERGIO_DS4_gene_expression_matrix,
                        ground_truth = SERGIO_DS4_ground_truth),
          in_format = "data_obj")

# inputs from a config file
# create a config YAML file with absolute paths to data
yaml::write_yaml(list(gene_expression_matrix = system.file("extdata", "SERGIO_DS4_gene_expression_matrix.csv", package = "CICTv2", mustWork = TRUE),
                ground_truth = system.file("extdata", "SERGIO_DS4_ground_truth.csv", package = "CICTv2", mustWork = TRUE),
                gene_association_matrix = NULL,
                edge_features = NULL,
                model = NULL,
                model_assessment = NULL,
                predicted_edges = NULL),
           "/home/syz248/CICTv2/inst/extdata/SERGIO_DS4_config.yaml")
checkData(config_path = system.file("extdata", "SERGIO_DS4_config.yaml", package = "CICTv2", mustWork = TRUE), in_format = "config_file")
}
