% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkGEM}
\alias{checkGEM}
\title{Check validity of gene expression matrix}
\usage{
checkGEM(gem)
}
\arguments{
\item{gem}{The object that will be checked to determine whether it is a valid
gene expression matrix.}
}
\value{
A list with fields \code{valid}, \code{warning}, and \code{error}. \code{valid} is a
boolean that is \code{TRUE} if the input exists, is a matrix or DataFrame, and
has numeric data; it is otherwise \code{FALSE}. \code{warning} is a boolean with
value \code{FALSE}. \code{message} is a string that contains an explanation of why
the input triggered a warning, was valid, or was invalid.
}
\description{
Checks if the gene expression matrix exists, is a matrix or DataFrame, and
has numeric data.
}
\examples{
TODO
}
