% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkGT}
\alias{checkGT}
\title{Check validity of ground truth table}
\usage{
checkGT(gt)
}
\arguments{
\item{gt}{The object that will be checked to determine whether it is a valid
ground truth table.}
}
\value{
A list with fields \code{valid}, \code{warning}, and \code{error}. \code{valid} is a
boolean that is \code{TRUE} if the input exists, and contains the columns "src"
and "trgt"; it is otherwise \code{FALSE}. \code{warning} is a boolean that is \code{TRUE}
if the ground truth table does not exists, and otherwise \code{FALSE}. \code{message}
is a string that contains an explanation of why the input triggered a
warning, was valid, or was invalid.
}
\description{
Checks if the ground truth table exists and contains the columns needed for
model training and evaluation.
}
\examples{
TODO
}
