% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICTv2-package.R
\docType{package}
\name{CICTv2-package}
\alias{CICTv2}
\alias{CICTv2-package}
\title{CICTv2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
