% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictEdges.R
\name{predictEdges}
\alias{predictEdges}
\title{predictEdges}
\usage{
predictEdges(
  edge_features = NULL,
  ground_truth = NULL,
  in_format = "separate",
  url.preset.train = NA,
  url.preset.test = NA,
  learning_ratio = 0.8,
  maxGroundTruth = 500,
  randomEdgesFoldCausal = 5,
  negativeEdgesFoldCausal = 1,
  exportTrainAndTest = T,
  returnDat = T,
  include.negative = "random",
  remove.tfs = T,
  split.val.tfs = F,
  learning_params = NA,
  trainingTarget = "class2",
  tstPercent = 0.3,
  predict_on = "none",
  split_ground_truth_by = "nodes",
  sample_tfs = "random",
  remove_learning = TRUE,
  ...
)
}
\arguments{
\item{edge_features}{NEW NAME FOR dt.edge}

\item{ground_truth}{NEW NAME for tbl.goldStandard}

\item{learning_ratio}{percent of ground truth to be used for learning}

\item{randomEdgesFoldCausal}{Numeric representing the number of irrelevant edges to be included
in the learning set relative to the number of causal edges. Default is 5.}

\item{negativeEdgesFoldCausal}{Numeric representing the number of negative edges to be included
in the learning set relative to the number of causal edges. Default is 1.}

\item{predict_on}{Default is: 'none'. User can provide strings 'none', 'all', or a vector
of genes on which the user wants to predict.}

\item{sample_tfs}{Default is: 'random'. Used only when param split_ground_truth_by is set to 'tfs'.
User can provide strings 'quartiles' or 'random'. Option 'quartiles' will  randomly sample from each
quartile of TFs ranked by number of targets. 'random' will randomly sample from the entire list in an unbiased fashion.}

\item{remove_learning}{Removes the learning set from the final predictions for performance
evaluation. Prevents bias when learning set TFs will always have inflated predictive scores.
Default is TRUE.}

\item{dt.edge}{CICT edges produced by prepareEdgeFeatures}

\item{preset.train}{Defualt is: NA. If provided a path to proper CSV, uses that for training.
Useful for sensitivity analysis as well as comparision with other methods on similar set of edges/features}

\item{preset.test}{Defualt is: NA. If provided a path to proper CSV, uses that for training.
Useful for sensitivity analysis as well as comparision with other methods on similar set of edges/features}
}
\value{
Returns a list consisting of three objects
edges: a dataframe of edge objects and CICT features for edges
Vertices: a dataframe of vertices objects and CICT features for vertices
}
\description{
Implements CICT supervised learning and prediction of regulatory edges. Currently heavily depends on global variables
}
\examples{
out <- predictEdges(edge_features = SERGIO_DS4_net0_edge_features,
                    ground_truth = SERGIO_DS4_net0_ground_truth,
                    in_format = 'separate',
                    randomEdgesFoldCausal = 5,
                    learning_ratio = 0.7)
}
