% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareEdgeFeatures.R
\name{prepareEdgeFeatures}
\alias{prepareEdgeFeatures}
\title{Prepare Edge Features}
\usage{
prepareEdgeFeatures(
  raw_edges = NULL,
  gene_expression_matrix = NULL,
  cict_raw_edge_col = "Spearman",
  in_format = "separate",
  prior = NULL,
  ...
)
}
\arguments{
\item{raw_edges}{Data frame. Raw edges data. Default is NULL.}

\item{gene_expression_matrix}{Data frame. Gene expression matrix. Default is NULL.}

\item{cict_raw_edge_col}{Character. Column name for raw edge calculation. Default is 'Spearman'.}

\item{in_format}{Character. Format of the input data. Can be "separate" or "data_obj". Default is "separate".}

\item{...}{Additional arguments to be passed to other functions.}
}
\value{
A list containing the prepared edge features and other relevant data.
}
\description{
This function prepares edge features from raw edges and a gene expression matrix.
}
\details{
This function processes the raw edges and gene expression matrix to prepare edge features. It supports two input formats: "separate" and "data_obj".
}
\examples{
\dontrun{
# Example usage:
raw_edges <- read.csv("path/to/raw_edges.csv")
gene_expression_matrix <- read.csv("path/to/gene_expression_matrix.csv")
edge_features <- prepareEdgeFeatures(
  raw_edges = raw_edges,
  gene_expression_matrix = gene_expression_matrix,
  cict_raw_edge_col = 'Spearman'
)
}
}
