% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SERGIO_DS4_ground_truth}
\alias{SERGIO_DS4_ground_truth}
\title{SERGIO DS4 ground truth}
\format{
\subsection{\code{SERGIO_DS4_ground_truth}}{

A data frame with 137 rows and 2 columns:
\describe{
\item{src}{Source gene}
\item{trgt}{Target gene}
}
}
}
\source{
\url{https://github.com/hlab1/scRNAseqWithCICT/blob/main/inputs/SERGIO_DS4/net0/refNetwork.csv}
}
\usage{
SERGIO_DS4_ground_truth
}
\description{
A ground truth gene regulatory network for the SERGIO DS4 generated gene expression data
}
\keyword{datasets}
