% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SERGIO_DS4_net0_gene_expression_matrix}
\alias{SERGIO_DS4_net0_gene_expression_matrix}
\title{SERGIO DS4 gene expression matrix}
\format{
\subsection{\code{SERGIO_DS4_net0_gene_expression_matrix}}{

A gzipped data frame with 100 rows and 900 columns:
\describe{
\item{Rows}{Unique gene names}
\item{Columns}{Unique sample names}
\item{Values}{Expression levels of each gene in each sample}
}
}
}
\source{
\url{https://github.com/hlab1/scRNAseqWithCICT/blob/main/inputs/SERGIO_DS4/net0/ExpressionData.csv}
}
\usage{
SERGIO_DS4_net0_gene_expression_matrix
}
\description{
The DS4 gene expression matrix generated with SERGIO
}
\keyword{datasets}
