% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRawEdges.R
\name{calculateRawEdges}
\alias{calculateRawEdges}
\title{Calculate Raw Edges}
\usage{
calculateRawEdges(
  n.workers = 5,
  in_data_obj = NULL,
  raw_edges = NULL,
  gene_expression_matrix = NULL,
  cict_raw_edge_col = "Spearman",
  in_format = "separate",
  ...
)
}
\arguments{
\item{n.workers}{Integer. Number of workers for parallel processing. Default is 5.}

\item{in_data_obj}{List. Input data object containing necessary data.}

\item{raw_edges}{Data frame. Raw edges data.}

\item{gene_expression_matrix}{Data frame. Gene expression matrix.}

\item{cict_raw_edge_col}{Character. Column name for raw edge calculation. Default is 'Spearman'.}

\item{in_format}{Character. Format of the input data. Default is "separate".}
}
\value{
A list containing the calculated raw edges and other relevant data.
}
\description{
This function calculates raw edges using mutual information and other measures in parallel.
}
\details{
This function performs parallel processing to calculate raw edges using mutual information and other measures. It is a modified edition of the code provided by Kuzmanovski et al.
}
\examples{
\dontrun{
# Example usage:
gene_expression_matrix <- read.csv("path/to/gene_expression_matrix.csv")
raw_edges <- read.csv("path/to/raw_edges.csv")
results <- calculateRawEdges(
  n.workers = 5,
  gene_expression_matrix = gene_expression_matrix,
  raw_edges = raw_edges,
  cict_raw_edge_col = 'Spearman'
)
}
}
