% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictEdges.R
\name{PredictEdges}
\alias{PredictEdges}
\title{PredictEdges}
\usage{
PredictEdges(
  edge_features = NULL,
  gene_expression_matrix = NULL,
  ground_truth = NULL,
  in_data_obj = NULL,
  in_format = "separate",
  method = "rf",
  evaluateUnseenSample = T,
  evaluateAllEdges = F,
  Debug = F,
  url.preset.train = NA,
  url.preset.test = NA,
  minGroundTruth.ratio.learning = 0.3,
  maxGroundTruth = 500,
  randomEdgesFoldCausal = 5,
  RF_ntrees = 50,
  RF_max_depth = 20,
  exportTrainAndTest = T,
  returnDat = T,
  include.negative = "random",
  rep = "",
  remove.tfs = T,
  split.val.tfs = F,
  learning.params = NA,
  runOnAllEdges = T,
  trainingTarget = "class2",
  tstPercent = 0.3,
  url.logfile = "noLog"
)
}
\arguments{
\item{edge_features}{NEW NAME FOR dt.edge}

\item{gene_expression_matrix}{gene expression matrix not required for this function}

\item{ground_truth}{NEW NAME for tbl.goldStandard}

\item{in_data_obj}{cict object replacing rcrd that has all of the results stored as a list}

\item{method}{Default value is: "CICT", which uses CICT features in a supervised learning. Also can take "RF" (random forest) or "XGB" (xgboost) which these supervised methods on relevance measures}

\item{evaluateUnseenSample}{Defualt is: TRUE, calculates regulatory edges on a subsample of up to 50000 edges.}

\item{evaluateAllEdges}{Defualt is: FALSE. If TRUE, CICT tries predicting all potential edges in the given network which might take longer time.}

\item{Debug}{If TRUE, function enters debugging mode in critical stops along the exectuion allowing verification of variables and features}

\item{dt.edge}{CICT edges produced by prepareEdgeFeatures}

\item{rcrd}{A list object that accumulates intermediary objects, results and performance measures and will be stored as an rds file for later use}

\item{preset.train}{Defualt is: NA. If provided a path to proper CSV, uses that for training. Useful for sensitivity analysis as well as comparision with other methods on similar set of edges/features}

\item{preset.test}{Defualt is: NA. If provided a path to proper CSV, uses that for training. Useful for sensitivity analysis as well as comparision with other methods on similar set of edges/features}
}
\value{
Returns a list consisted of three objects
rcrd: is a list object of intermediary objects
edges: a dataframe of edge objects and CICT features for edges
Vertices: a dataframe of vertices objects and CICT features for vertices
}
\description{
Implements CICT supervised learning and prediction of regulatory edges. Currently heavily depends on global variables
}
\examples{
c(rcrd,edges,vertices) \%<-z\% prepareEdgeFeatures(Debug=Debug)
}
