% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictEdges.R
\name{predictEdges}
\alias{predictEdges}
\title{predictEdges}
\usage{
predictEdges(
  edge_features = NULL,
  ground_truth = NULL,
  learning_ratio = 0.8,
  randomEdgesFoldCausal = 5,
  negativeEdgesFoldCausal = 1,
  exportTrainAndTest = T,
  returnDat = T,
  include.negative = "random",
  remove.tfs = T,
  split.val.tfs = F,
  learning_params = NA,
  trainingTarget = "class2",
  tstPercent = 0.3,
  predict_on = "none",
  split_ground_truth_by = "nodes",
  sample_tfs = "random",
  remove_learning = TRUE,
  ...
)
}
\arguments{
\item{edge_features}{CICT edges produced by prepareEdgeFeatures}

\item{ground_truth}{Ground truth table}

\item{learning_ratio}{percent of ground truth to be used for learning}

\item{randomEdgesFoldCausal}{Numeric representing the number of irrelevant edges to be included
in the learning set relative to the number of causal edges. Default is 5.}

\item{negativeEdgesFoldCausal}{Numeric representing the number of negative edges to be included
in the learning set relative to the number of causal edges. Default is 1.}

\item{predict_on}{Default is: 'none'. User can provide strings 'none', 'all', or a vector
of genes on which the user wants to predict.}

\item{sample_tfs}{Default is: 'random'. Used only when param split_ground_truth_by is set to 'tfs'.
User can provide strings 'quartiles' or 'random'. Option 'quartiles' will  randomly sample from each
quartile of TFs ranked by number of targets. 'random' will randomly sample from the entire list in an unbiased fashion.}

\item{remove_learning}{Removes the learning set from the final predictions for performance
evaluation. Prevents bias when learning set TFs will always have inflated predictive scores.
Default is TRUE.}

\item{preset.train}{Defualt is: NA. If provided a path to proper CSV, uses that for training.
Useful for sensitivity analysis as well as comparision with other methods on similar set of edges/features}

\item{preset.test}{Defualt is: NA. If provided a path to proper CSV, uses that for training.
Useful for sensitivity analysis as well as comparision with other methods on similar set of edges/features}
}
\value{
A list in the CICT data object format. Contains \code{model},
\code{model_assessment}, and \code{predicted_edges}.
}
\description{
Implements CICT supervised learning and prediction of regulatory edges. Currently heavily depends on global variables
}
\examples{
# Download data from the external data folder of the CICTv2 GitHub repo
download.file("https://raw.githubusercontent.com/hlab1/CICTv2/refs/heads/main/inst/extdata/SERGIO_DS4_net0_gene_expression_matrix.csv",
"SERGIO_DS4_net0_gene_expression_matrix.csv")
download.file("https://raw.githubusercontent.com/hlab1/CICTv2/refs/heads/main/inst/extdata/SERGIO_DS4_net0_ground_truth.csv",
"SERGIO_DS4_net0_ground_truth.csv")
download.file("https://raw.githubusercontent.com/hlab1/CICTv2/refs/heads/main/inst/extdata/SERGIO_DS4_net0_edge_features.csv",
"SERGIO_DS4_net0_edge_features.csv")

gene_expression_matrix <- read.csv(
  "SERGIO_DS4_net0_gene_expression_matrix.csv",
  header = TRUE,
  row.names = 1)
edge_features <- read.table("SERGIO_DS4_net0_edge_features.csv",
  header=TRUE,
  sep = ",")
ground_truth <- read.table("SERGIO_DS4_net0_ground_truth.csv",
  header=TRUE,
  sep = ",")
predictEdges(gene_expression_matrix = gene_expression_matrix,
  ground_truth = ground_truth,
  edge_features = edge_features)

# Reset workspace
unlink("SERGIO_DS4_net0_gene_expression_matrix.csv")
unlink("SERGIO_DS4_net0_edge_features.csv")
unlink("SERGIO_DS4_net0_ground_truth.csv")
rm(gene_expression_matrix)
rm(edge_features)
rm(ground_truth)
}
