% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareEdgeFeatures.R
\name{prepareEdgeFeatures}
\alias{prepareEdgeFeatures}
\title{Prepare Edge Features}
\usage{
prepareEdgeFeatures(
  raw_edges = NULL,
  gene_expression_matrix = NULL,
  cict_raw_edge_col = "Spearman",
  prior = NULL,
  ...
)
}
\arguments{
\item{raw_edges}{Data frame. Raw edges data. Default is NULL.}

\item{gene_expression_matrix}{Data frame. Gene expression matrix. Default is NULL.}

\item{cict_raw_edge_col}{Character. Column name for raw edge calculation. Default is 'Spearman'.}

\item{...}{Additional arguments to be passed to other functions.}
}
\value{
A list containing the prepared edge features and other relevant data.
}
\description{
This function prepares edge features from raw edges and a gene expression matrix.
}
\details{
This function processes the raw edges and gene expression matrix to prepare edge features. It supports two input formats: "separate" and "data_obj".
}
\examples{
# Download data from the external data folder of the CICTv2 GitHub repo
download.file("https://raw.githubusercontent.com/hlab1/CICTv2/refs/heads/main/inst/extdata/SERGIO_DS4_net0_gene_expression_matrix.csv",
"SERGIO_DS4_net0_gene_expression_matrix.csv")
download.file("https://raw.githubusercontent.com/hlab1/CICTv2/refs/heads/main/inst/extdata/SERGIO_DS4_net0_raw_edges.csv",
"SERGIO_DS4_net0_raw_edges.csv")
gene_expression_matrix <- read.csv("SERGIO_DS4_net0_gene_expression_matrix.csv", header = TRUE, row.names = 1)
raw_edges <- read.table("SERGIO_DS4_net0_raw_edges.csv",  header=TRUE, sep = ",")
prepareEdgeFeatures(gene_expression_matrix = gene_expression_matrix,
        raw_edges = raw_edges, cict_raw_edge_col = "Pearson")

# Reset workspace
unlink("SERGIO_DS4_net0_gene_expression_matrix.csv")
unlink("SERGIO_DS4_net0_raw_edges.csv")
rm(gene_expression_matrix)
rm(raw_edges)
}
