% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCICT.R
\name{runCICT}
\alias{runCICT}
\title{Driver for the CICT pipeline}
\usage{
runCICT(
  gene_expression_matrix = NULL,
  ground_truth = NULL,
  config_path = NULL,
  in_format = "separate",
  ...
)
}
\arguments{
\item{gene_expression_matrix}{The gene expression matrix, where each row
represents a gene and each column represents a sample. A matrix or Data
Frame with numeric data.}

\item{ground_truth}{A Data Frame representing the ground-truth gene
regulatory network for model training and evaluation. Each row represents a
source-target relationship, with the source gene in the column labeled
\code{"src"} and the target gene in the column labeled \code{"trgt"}}

\item{config_path}{Path to the YAML config file, if one is used.}

\item{in_format}{String indicating expected input format. \code{"separate"} if
passing inputs through \code{gene_expression_matrix} and \code{ground_truth},
\code{"config_file"} if passing inputs through \code{config_path}. If using the config
format, all other arguments passed to the function are ignored except
\code{config_path}, and the resulting predicted edges and CICT data object  will
be stored as a CSV and .Rds file respectively. Use "results_dir" argument
in config file to specify where these files should be saved; otherwise,
they will be saved to the working directory.}

\item{...}{Options to be passed to calculateRawEdges, prepareEdgeFeatures,
and/or predictEdges}
}
\value{
A list in the CICT data object format, with the data from each step
in the CICT pipeline. Contains \code{gene_expression_matrix}, \code{ground_truth},
\code{raw_edges}, \code{edge_features}, \code{model}, \code{model_assessment},
\code{predicted_edges}, and potentially other data.
}
\description{
Takes a gene expression matrix and a ground truth table, calculates raw
edges, prepares edge features, and uses a random forest model to predict a
gene regulatory network. If in_format is specified as config_path, inputs and
arguments will come from a YAML config file.
}
\examples{
# Download data from the external data folder of the CICTv2 GitHub repo
download.file("https://raw.githubusercontent.com/hlab1/CICTv2/refs/heads/main/inst/extdata/SERGIO_DS4_net0_gene_expression_matrix.csv", "SERGIO_DS4_net0_gene_expression_matrix.csv")
download.file("https://raw.githubusercontent.com/hlab1/CICTv2/refs/heads/main/inst/extdata/SERGIO_DS4_net0_ground_truth.csv", "SERGIO_DS4_net0_ground_truth.csv")
download.file("https://raw.githubusercontent.com/hlab1/CICTv2/refs/heads/main/inst/extdata/SERGIO_DS4_net0_config.yaml", "SERGIO_DS4_net0_config.yaml")
# Separate inputs mode:
gene_expression_matrix <- read.csv("SERGIO_DS4_net0_gene_expression_matrix.csv", header = TRUE, row.names = 1)
ground_truth <- read.table("SERGIO_DS4_net0_ground_truth.csv",  header=TRUE, sep = ",")
runCICT(gene_expression_matrix = gene_expression_matrix,
        ground_truth = ground_truth)

# Config mode:
runCICT(config_path = "SERGIO_DS4_net0_config.yaml", in_format = "config_file")

# Reset workspace
unlink("SERGIO_DS4_net0_gene_expression_matrix.csv")
unlink("SERGIO_DS4_net0_ground_truth.csv")
unlink("SERGIO_DS4_net0_config.yaml")
unlink("SERGIO_DS4_net0/log")
unlink("SERGIO_DS4_net0/predicted_edges.csv")
unlink("SERGIO_DS4_net0/cict_data.Rds")
rm(gene_expression_matrix)
rm(ground_truth)
}
